feat: `/calendar` route → `CalendarPage`
[claude]
