+++ src/pages/NotFoundPage.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { useHead } from '@unhead/vue'
+import FooterBar from '@/components/FooterBar.vue'
 
 useHead({ title: 'Page Not Found — King Gizzard History' })
 </script>
@@ -9,6 +10,7 @@ useHead({ title: 'Page Not Found — King Gizzard History' })
     <h1>Page not found</h1>
     <p><RouterLink to="/">Back to Calendar</RouterLink></p>
   </main>
+  <FooterBar />
 </template>
 
 <style scoped>
